+++ 1.20.1/common/src/main/java/com/github/hoshinofw/buildstonetoolkit/foundation/client/util/ParticleUtil.java
@@ -1,10 +1,12 @@
-package com.github.hoshinofw.buildstonetoolkit.foundation.util;
+package com.github.hoshinofw.buildstonetoolkit.foundation.client.util;
 
 import com.github.hoshinofw.buildstonetoolkit.foundation.common.blocks.IdProxyBlock;
 import com.github.hoshinofw.buildstonetoolkit.foundation.common.blocks.ProxyBlock;
 import com.github.hoshinofw.buildstonetoolkit.foundation.common.blocks.entity.unstable.IdProxyBlockEntity;
 import com.github.hoshinofw.buildstonetoolkit.foundation.client.particles.StaticCubeParticle.CubeParticle;
 import com.github.hoshinofw.buildstonetoolkit.foundation.common.registries.BuildstoneItems;
+import com.github.hoshinofw.buildstonetoolkit.foundation.util.BlockPosSupplier;
+import com.github.hoshinofw.buildstonetoolkit.foundation.util.PlayerUtil;
 import com.github.hoshinofw.buildstonetoolkit.foundation.util.render.BlockParticleTexture;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;